requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.2.6
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
requests==2.31.0
tqdm==4.66.1
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
